requirements.txt
# FastAPI и веб-сервер
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# База данных
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0
greenlet>=3.0.0

# Аутентификация
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Валидация и настройки
pydantic==2.10.4
pydantic-settings==2.7.0

# Telegram Bot
aiogram==3.16.0
httpx==0.28.1

# Production
gunicorn==23.0.0
